+++ server/config/config.go
@@ -15,6 +15,7 @@ import (
 	errorsmod "cosmossdk.io/errors"
 	"github.com/cosmos/cosmos-sdk/server/config"
 	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
+	"github.com/crypto-org-chain/cronos/memiavl"
 	memiavlcfg "github.com/crypto-org-chain/cronos/store/config"
 )
 
@@ -93,6 +94,27 @@ const (
 
 	// DefaultGasAdjustment value to use as default in gas-adjustment flag
 	DefaultGasAdjustment = 1.2
+
+	// ============================
+	//           MemIAVL
+	// ============================
+
+	// DefaultMemIAVLEnable is the default value that defines if memIAVL is enabled
+	DefaultMemIAVLEnable = false
+
+	// DefaultZeroCopy is the default value that defines if
+	// the zero-copied slices must be retained beyond current block's execution
+	// the sdk address cache will be disabled if zero-copy is enabled
+	DefaultZeroCopy = false
+
+	// DefaultAsyncCommitBuffer value to use as default for the size of
+	// asynchronous commit queue when using memIAVL
+	DefaultAsyncCommitBuffer = 0
+
+	// DefaultSnapshotKeepRecent default value for how many old snapshots
+	// (excluding the latest one) should be kept after new snapshots
+	// when using memIAVL
+	DefaultSnapshotKeepRecent = 1
 )
 
 var evmTracers = []string{"json", "markdown", "struct", "access_list"}
@@ -348,7 +370,14 @@ func (c TLSConfig) Validate() error {
 
 // DefaultMemIAVLConfig returns the default MemIAVL configuration
 func DefaultMemIAVLConfig() *MemIAVLConfig {
-	return &MemIAVLConfig{memiavlcfg.DefaultMemIAVLConfig()}
+	return &MemIAVLConfig{memiavlcfg.MemIAVLConfig{
+		Enable:             DefaultMemIAVLEnable,
+		ZeroCopy:           DefaultZeroCopy,
+		AsyncCommitBuffer:  DefaultAsyncCommitBuffer,
+		SnapshotKeepRecent: DefaultSnapshotKeepRecent,
+		SnapshotInterval:   memiavl.DefaultSnapshotInterval,
+		CacheSize:          memiavlcfg.DefaultCacheSize,
+	}}
 }
 
 // Validate returns an error if the MemIAVL configuration fields are invalid.