+++ utils/utils.go
@@ -4,7 +4,10 @@
 package utils
 
 import (
+	"fmt"
 	"strings"
+
+	ibctransfertypes "github.com/cosmos/ibc-go/v7/modules/apps/transfer/types"
 
 	"github.com/evmos/evmos/v15/crypto/ethsecp256k1"
 
@@ -84,4 +87,37 @@ func GetEvmosAddressFromBech32(address string) (sdk.AccAddress, error) {
 	}
 
 	return sdk.AccAddress(addressBz), nil
+}
+
+// CreateAccAddressFromBech32 creates an AccAddress from a Bech32 string.
+func CreateAccAddressFromBech32(address string, bech32prefix string) (addr sdk.AccAddress, err error) {
+	if len(strings.TrimSpace(address)) == 0 {
+		return sdk.AccAddress{}, fmt.Errorf("empty address string is not allowed")
+	}
+
+	bz, err := sdk.GetFromBech32(address, bech32prefix)
+	if err != nil {
+		return nil, err
+	}
+
+	err = sdk.VerifyAddressFormat(bz)
+	if err != nil {
+		return nil, err
+	}
+
+	return sdk.AccAddress(bz), nil
+}
+
+// ComputeIBCDenom compute the ibc voucher denom associated to
+// the portID, channelID, and the given a token denomination.
+func ComputeIBCDenom(
+	portID, channelID,
+	denom string,
+) string {
+	denomTrace := ibctransfertypes.DenomTrace{
+		Path:      fmt.Sprintf("%s/%s", portID, channelID),
+		BaseDenom: denom,
+	}
+
+	return denomTrace.IBCDenom()
 }